+++ tests/base.test.js
@@ -704,7 +704,7 @@ testTree.getNodeByColumns = equal => {
  description Test a root parser node
  constructors
   nodejs ./JibberishLang.js JibberishProgramRoot
- compiler txt
+ compiler
  inScope baseNode`)
 
   // Act
@@ -1900,7 +1900,7 @@ testTree.isomorphicGrammarTests = equal => {
  description Test a root parser node
  constructors
   nodejs JibberishLang.js
- compiler txt
+ compilesTo txt
  catchAllNodeType error
  inScope topLevel text someAbstractClass
 cellType int
@@ -1949,7 +1949,7 @@ nodeType lightbulbState
 nodeType to
  extends block
  cells word
- compiler txt
+ compiler
   stringTemplate to {word}
   closeChildren end`
   const code = `foo